requirements.txt
boto3==1.34.27
botocore==1.34.27
certifi==2023.11.17
charset-normalizer==3.3.2
datadog==0.48.0
idna==3.6
jmespath==1.0.1
mysql-connector-python==8.3.0
pycryptodome==3.20.0
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
urllib3==1.26.18
